% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bep.R
\name{bep_dollars}
\alias{bep_dollars}
\title{Break even point dollars}
\usage{
bep_dollars(variable_cost = "", fixed_cost = "", price = "")
}
\arguments{
\item{variable_cost}{}

\item{fixed_cost}{}

\item{price}{}
}
\value{

}
\description{
Calculate BEP in dollars
}
