% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bep.R
\name{bep_units}
\alias{bep_units}
\title{Break even point units}
\usage{
bep_units(variable_cost = "", fixed_cost = "", price = "")
}
\arguments{
\item{variable_cost}{}

\item{fixed_cost}{}

\item{price}{}
}
\value{

}
\description{
Calculate BEP in units
}
